+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [0.4.0] - 2021-11-01
+
+### Added
+
+- Switch checkbox component.
+- Some constants to exports.
+
 ## [0.3.0] - 2021-10-30
 
 ### Changed
@@ -45,7 +52,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 - Card component.
 - Material icon component.
 
-[unreleased]: https://github.com/visiosto/react-components/compare/v0.3.0...HEAD
+[unreleased]: https://github.com/visiosto/react-components/compare/v0.4.0...HEAD
+[0.4.0]: https://github.com/visiosto/react-components/compare/v0.3.0...v0.4.0
 [0.3.0]: https://github.com/visiosto/react-components/compare/v0.2.1...v0.3.0
 [0.2.1]: https://github.com/visiosto/react-components/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/visiosto/react-components/compare/v0.1.2...v0.2.0